web-ui/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0